--- Jason/src/routes/jobs.routes.ts
@@ -1,8 +1,0 @@
-import { Router } from "express";
-import { getJobsController } from "../controllers/jobs.controller";
-
-const router = Router();
-
-router.get("/", getJobsController);
-
-export default router;
